Replacing deprecated MPI 1.0 functions

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@d7e6f7f235e09b15e64baa410fb9cc2e47e9e0ce

+++ src/coreneuron/nrniv/ivlistimpl.cpp
@@ -41,7 +41,7 @@ static long ListImpl_best_new_sizes[] = {
 
 long ListImpl_best_new_count(long count, unsigned int size, unsigned int m) {
     for (size_t i = 0; i < sizeof(ListImpl_best_new_sizes)/sizeof(long); i++) {
-        if (count * size < ListImpl_best_new_sizes[i]) {
+        if ( (long(count * size)) < ListImpl_best_new_sizes[i]) {
             return ListImpl_best_new_sizes[i] / size;
         }
     }
